requirements.txt
--find-links https://download.pytorch.org/whl/torch_stable.html
torch==2.6.0+cpu
torchvision==0.21.0+cpu
numpy>=1.24.3
matplotlib>=3.7.1
Pillow>=9.5.0
tqdm>=4.65.0
streamlit>=1.22.0
scikit-learn>=1.2.2
pandas>=2.0.1
opencv-python 